Make the routes utility functions accessible globally
